% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_fact_det}
\alias{fast_fact_det}
\title{Determinant of a matrix using low-rank plus diagonal decomposition}
\usage{
fast_fact_det(Lambda_orig, ps, lg = 1L)
}
\arguments{
\item{Lambda_orig}{A \eqn{d \times q} order matrix \eqn{\Lambda} typically \eqn{q\ll d}}

\item{ps}{A vector of length \eqn{d} comprising positive elements}

\item{lg}{A logical variable; if \code{TRUE} (default), \code{log} of the determinant will be returned.}
}
\value{
If \code{lg=TRUE} then \code{log}-determinant of \eqn{\Lambda \Lambda^T+\mathrm{diag}(\texttt{ps}) }; else determinant of \eqn{\Lambda \Lambda^T+\mathrm{diag}(\texttt{ps}) }.
}
\description{
Fast determinant of \eqn{\Sigma=\Lambda \Lambda^T+\mathrm{diag}(\texttt{ps})} using the \href{https://en.wikipedia.org/wiki/Matrix_determinant_lemma}{Matrix determinant lemma}
}
\examples{
d=1e3; q=50
lam=matrix(rnorm(d*q,sd=.25),nrow=d,ncol=q ); ps=rgamma(d,1,1) 
sig=tcrossprod(lam)+diag(ps)
rbenchmark::benchmark( log(det(sig)), ##naive determinant
2*sum(log(diag(chol(sig ))) ),##Cholesky based determinant
 fast_fact_det(lam,ps), ##fast_fact_det
 replications = 10  ##benchmark using 10 replicates
 ) 
}
