% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_fact_inv}
\alias{fast_fact_inv}
\title{Matrix inversion using low-rank plus diagonal decomposition}
\usage{
fast_fact_inv(Lambda, ps)
}
\arguments{
\item{Lambda}{A \eqn{d \times q} order matrix \eqn{\Lambda} typically \eqn{q\ll d}}

\item{ps}{A vector of length \eqn{d} comprising positive elements}
}
\value{
The matrix \eqn{(\Lambda \Lambda^T+\mathrm{diag}(\texttt{ps}) )^{-1}}
}
\description{
Fast inversion of \eqn{\Sigma=\Lambda \Lambda^T+\mathrm{diag}(\texttt{ps})} using the \href{https://en.wikipedia.org/wiki/Woodbury_matrix_identity}{Woodbury matrix identity}
}
\examples{
d=1e3; q=50
lam=matrix(rnorm(d*q,sd=.25),nrow=d,ncol=q ); ps=rgamma(d,1,1) 
sig=tcrossprod(lam)+diag(ps)
rbenchmark::benchmark( chol2inv(chol(sig )),##Cholesky based inverse
 fast_fact_inv(lam,ps), ##fast_fact_inv
 replications = 10  ##benchmark using 10 replicates
 ) 
}
