% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{fit_FA}
\alias{fit_FA}
\title{Fit a Bayesian sparse factor analysis model}
\usage{
fit_FA(
  Y,
  qmax,
  a = 0.5,
  ms = 1,
  ss = 7,
  nrun = 7500,
  thin = 5,
  nleapfrog = 6,
  leapmax = 10,
  del_range = c(0.001, 0.0075)
)
}
\arguments{
\item{Y}{Data matrix}

\item{qmax}{Maximum allowed column dimension of the factor loading matrix.}

\item{a}{Dirichlet-Laplace shrinkage parameter.}

\item{ms}{Prior expectation of the idiosyncratic variance parameters \eqn{\delta_{j}^{2}}'s for all \eqn{j=1,\dots,d}.}

\item{ss}{Prior variance of the idiosyncratic variance parameters \eqn{\delta_{j}^{2}}'s for all \eqn{j=1,\dots,d}.}

\item{nrun}{Number of MCMC iterations to perform.}

\item{thin}{Every \code{thin}-th MCMC sample is stored.}

\item{nleapfrog}{The number of leapfrog steps in each MCMC iteration is sampled from \eqn{\min\{1,\text{Poisson}(\texttt{nleapfrog}) \} }.}

\item{leapmax}{The number of leapfrog steps is bounded above by \code{leapmax}.}

\item{del_range}{The step size of \eqn{\delta t} in  each HMC step is simulated as \eqn{ \delta t \sim} Uniform\code{(del_range[1],del_range[2])}.}
}
\value{
A list with the following elements:
\describe{
\item{\code{Lambda}}{An array with the MCMC samples of \eqn{\mathbf{\Lambda}}. MCMC samples are stored along the third dimension.}
\item{\code{residuals}}{A matrix with the MCMC samples of \eqn{\mathrm{diag}(\mathbf{\Delta})=(\delta_{1}^{2},\dots,\delta_{d}^{2})^{T}}. The MCMC samples are stored along the rows of \code{residuals}.}
\item{\code{acceptance_probability}}{The acceptance probability of the HMC-within-Gibbs sampler.}
}
}
\description{
A fast Hamiltonian Monte Carlo-based sparse Bayesian factor model implementation
\deqn{Y_{i}\sim \mathrm{N} (0, \mathbf{\Lambda}\mathbf{\Lambda}^{T}+\mathbf{\Delta}) \text{ where } \mathrm{diag}(\mathbf{\Delta})=(\delta_{1}^{2},\dots,\delta_{d}^{2})^{T}}
\deqn{\mathrm{vectorise}(\mathbf{\Lambda})\sim \mathrm{Dirichlet-Laplace}(a), }
\deqn{\log\delta_{j}^{2} \sim \mathrm{N}(\mu_{\delta},\sigma_{\delta}^{2}), } such that \eqn{\mu_{\delta}} and \eqn{\sigma_{\delta}^{2}} are chosen such that a priori \eqn{E(\delta_{j}^{2})=}\code{ms} and \eqn{\mathrm{var}(\delta_{j}^{2})=}\code{ss} for all \eqn{j=1,\dots,d}.
}
\examples{
vignette(topic="sparse_BFA",package = "SUFA")
}
