% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{fit_SUFA}
\alias{fit_SUFA}
\title{Fit the SUFA model}
\usage{
fit_SUFA(
  Y,
  qmax,
  qs = NULL,
  a = 0.5,
  bAs = 1,
  ms = 1,
  ss = 7,
  nthreads = 6,
  nrun = 7500,
  thin = 5,
  nleapfrog = 6,
  leapmax = 10,
  del_range = c(0.001, 0.0075),
  col_prop = 1
)
}
\arguments{
\item{Y}{A List of datasets. In each dataset independent samples MUST be stored along the rows. All of the datasets MUST comprise the same features.}

\item{qmax}{Maximum allowed column dimension of the shared factor loading matrix.}

\item{qs}{Vector of the number of study-specific latent factors. Default is \code{NULL} where the script will set \eqn{q_{s}=\frac{q}{S}} for all studies \eqn{s=1,\dots,S}.}

\item{a}{Dirichlet-Laplace shrinkage parameter.}

\item{bAs}{Prior variance hyperparameter of the elements in the study-specific \eqn{\mathbf{A}_{s}} matrices.}

\item{ms}{Prior expectation of the idiosyncratic variance parameters \eqn{\delta_{j}^{2}}'s for all \eqn{j=1,\dots,d}.}

\item{ss}{Prior variance of the idiosyncratic variance parameters \eqn{\delta_{j}^{2}}'s for all \eqn{j=1,\dots,d}.}

\item{nthreads}{Number of parallel threads to use.}

\item{nrun}{Number of MCMC iterations to perform.}

\item{thin}{Every \code{thin}-th MCMC sample is stored.}

\item{nleapfrog}{The number of leapfrog steps in each MCMC iteration is sampled from \eqn{\min\{1,\text{Poisson}(\texttt{nleapfrog}) \} }.}

\item{leapmax}{The number of leapfrog steps is bounded above by \code{leapmax}.}

\item{del_range}{The step size of \eqn{\delta t} in  each HMC step is simulated as \eqn{ \delta t \sim} Uniform\code{(del_range[1],del_range[2])}.}

\item{col_prop}{The proportion of randomly selected columns of \eqn{\mathbf{\Lambda}} to update in each HMC step. If the dimension of the data is high, we suggest lowering this value from 1.}
}
\value{
A list with the following elements:
\describe{
\item{\code{Lambda}}{An array with the MCMC samples of \eqn{\mathbf{\Lambda}}. MCMC samples are stored along the third dimension.}
\item{\code{A}}{A list of the MCMC samples of the study-specific \eqn{\mathbf{A}_s} matrices. The \eqn{s}-th element of the list corresponds to \eqn{\mathbf{A}_s} with MCMC samples stored in the same manner as that of \eqn{\mathbf{\Lambda}}.}
\item{\code{residuals}}{A matrix with the MCMC samples of \eqn{\mathrm{diag}(\mathbf{\Delta})=(\delta_{1}^{2},\dots,\delta_{d}^{2})^{T}}. The MCMC samples are stored along the rows of \code{residuals}.}
\item{\code{acceptance_probability}}{The acceptance probability of the HMC-within-Gibbs sampler.}
}
}
\description{
Fit the SUFA model
}
\examples{
vignette(topic="Intro_simulation",package = "SUFA")
}
