% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{heatplot}
\alias{heatplot}
\title{Heatmap of a matrix}
\usage{
heatplot(x)
}
\arguments{
\item{x}{A matrix}
}
\value{
\code{ggplot} object representing the heatmap of \code{x}
}
\description{
Heatmap of a matrix
}
\examples{
x=matrix(rnorm(50*5),nrow=50)
x[sample.int(length(x), 100 )]=NA
heatplot(x)
}
