% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{lam.est.all}
\alias{lam.est.all}
\title{Estimate shared and study-specific loading matrices from the output of \code{\link{fit_SUFA}}}
\usage{
lam.est.all(res, burn = 500, alpha = 0.05)
}
\arguments{
\item{res}{Output of \code{\link{fit_SUFA}}.}

\item{burn}{Number of burnin samples to discard.}

\item{alpha}{The elements of the loading matrices are coded as 0 if their respective (1-\code{alpha})\% posterior credible intervals (after fixing for rotational ambiguity) include 0.}
}
\value{
A list with the following elements:
\describe{
\item{\code{Shared}}{Point estimate of the shared loading matrix \eqn{\mathbf{\Lambda}}.}
\item{\code{Study_specific}}{A list comprising the point estimates of the study specific loading matrices \eqn{\mathbf{\Lambda A}_s}.}
}
}
\description{
Estimate shared and study-specific loading matrices from the output of \code{\link{fit_SUFA}}
}
