% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{lam.est}
\alias{lam.est}
\title{Post-process MCMC samples of loading matrices to resole rotational ambiguity}
\usage{
lam.est(lam_array, burn = 500, alpha = 0.05)
}
\arguments{
\item{lam_array}{An array of MCMC samples of the loading matrix. The MCMC samples MUST be stored along the third dimension.}

\item{burn}{Number of burnin samples to discard.}

\item{alpha}{The elements of the loading matrices are coded as 0 if their respective 100(1-\code{alpha})\% posterior credible intervals (after fixing for rotational ambiguity) include 0.}
}
\value{
A point estimate matrix
}
\description{
The varimax-based post-processing method of \insertCite{poworoznek2021;textual}{SUFA}. This function can be generally applied to any MCMC-based factor analysis model.
}
\references{
\insertAllCited{}
}
