% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_det_pd}
\alias{log_det_pd}
\title{Cholesky factorization based fast determinant of positive defnite matrices}
\usage{
log_det_pd(X, lg = 1L)
}
\arguments{
\item{X}{A positive definite matrix}

\item{lg}{A logical variable; if \code{TRUE} (default), \code{log} of the determinant will be returned.}
}
\value{
If \code{lg=TRUE} then \code{log}-determinant of \code{X}; else determinant of \code{X}
}
\description{
Cholesky factorization based fast determinant of positive defnite matrices
}
