\name{SUFA-package}
\alias{SUFA-package}
\alias{SUFA}
\docType{package}
\title{
\packageTitle{SUFA}
}
\description{
\packageDescription{SUFA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SUFA}
\packageIndices{SUFA}
A fast \code{C++} based implementation of the SUbspace factor analysis model by  \insertCite{chandra2023sufa;textual}{SUFA}. The Hamiltonian Monte Carlo with Gibbs sampler described is implemented as decribed in the article.
}
\author{
\packageAuthor{SUFA}

Maintainer: \packageMaintainer{SUFA}
}
\references{
\insertAllCited{}
}
\keyword{ SUFA }
\examples{
## See the vignettes included in this package.
}
