% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{sufa_marginal_covs}
\alias{sufa_marginal_covs}
\title{Computes posterior means of the marginal covariance matrix from the output of \code{\link{fit_SUFA}}}
\usage{
sufa_marginal_covs(res, burn = 500)
}
\arguments{
\item{res}{Output of \code{\link{fit_SUFA}}.}

\item{burn}{Number of burnin samples to discard.}
}
\value{
An array with the marginal covariance matrices in the studies concatenated across the third dimension.
}
\description{
Computes posterior means of the marginal covariance matrix from the output of \code{\link{fit_SUFA}}
}
