% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{SUFA_shared_covmat}
\alias{SUFA_shared_covmat}
\title{Computes posterior mean of the shared covariance matrix across studies from the output of \code{\link{fit_SUFA}}}
\usage{
SUFA_shared_covmat(res, burn = 500)
}
\arguments{
\item{res}{Output of \code{\link{fit_SUFA}}.}

\item{burn}{Number of burnin samples to discard.}
}
\value{
A matrix representing the posterior mean of \eqn{\mathbf{\Sigma}=\mathbf{\Lambda}\mathbf{\Lambda}^{T}+\mathbf{\Delta}}.
}
\description{
Computes posterior mean of the shared covariance matrix across studies from the output of \code{\link{fit_SUFA}}
}
\details{
Computes posterior mean of the shared covariance matrix \eqn{\mathbf{\Sigma}=\mathbf{\Lambda}\mathbf{\Lambda}^{T}+\mathbf{\Delta}} across studies from the output of \code{\link{fit_SUFA}}.
}
