% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{WBIC}
\alias{WBIC}
\title{Compute the widely applicable Bayesian information criterion (WBIC)}
\usage{
WBIC(res, y, model = "SUFA", burn = 500, ncores = 1)
}
\arguments{
\item{res}{Output of \code{\link{fit_SUFA}}.}

\item{y}{List of datasets which was fed into \code{\link{fit_SUFA}}.}

\item{model}{The default value should be used.}

\item{burn}{Number of burnin samples to discard.}

\item{ncores}{Should be set at 1.}
}
\value{
The WBIC value calculated from the MCMC samples.
}
\description{
Computes the widely applicable Bayesian information criterion (WBIC) \insertCite{wbic2013}{SUFA} from the output of \code{\link{fit_SUFA}}
}
\references{
\insertAllCited{}
}
